+++ src/main/kotlin/com/tianea/fimo/web/auth/dto/ReissueRequest.kt
@@ -1,0 +1,6 @@
+package com.tianea.fimo.web.auth.dto
+
+class ReissueRequest(
+    val accessToken: String,
+    val refreshToken: String
+)
